Refactor: Flatten module hierarchies and simplify structures
- Flatten `src/errors/` directory into `src/errors.rs`.
- Merge `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs`.
- Merge `src/parser/recursion.rs` into `src/parser/mod.rs`.
- Update imports and references accordingly.
- Remove deleted files and directories.
